feat: Add edit button to poll list
- Added an edit button to each poll in the poll list component
- The edit button is represented by the FiEdit icon from the react-icons library
- Clicking the edit button will log the poll ID to the console

This change adds an edit button to the poll list component, allowing users to easily edit each poll. The edit button is represented by the FiEdit icon from the react-icons library. When the edit button is clicked, the poll ID will be logged to the console for debugging purposes. This enhancement improves the user experience by providing a convenient way to edit polls directly from the poll list.